Decouple metrics collection from scraping (#26)
This commit decouples the collecting of metrics from the scraping
perfomed by Prometheus. This is achieved by running a background
thread wich collects the metrics from Jira and stores it in a global
datastructure. When the root context is queried this datastructure is
simply returned. The background job is, as of now, configured to run
every 5 minutes.

The metrics removed in previous commits are reinstated.

Co-authored-by: @are3k
Co-authored-by: @ArielKarlsen