Introduce structured logging
This commit will introduce structured logging i json-format as the
logs will be collected in Kibana. Also added more logging statements
and created two simple metrics for this application regarding the
execution time and the total number of metrics collected for each
run.

(Still some refactoring to be done but this will be added in
later commits/PRs)